refactor: Rename module 'key' to 'context_key'.
